SXC Pressure based around a mermaid-necromancer controlling the lake
The idea is to have the boss able to send creeps across the
water, but for the players to have to either lower the water
several times, or to lower it once and then attack from the
north side of the central castle. Either method would bring
the east or mid-west parts of the map in to play, working
towards the north-east or north-west boss.

This brings a scenario from the unbalanced directory back in to play.