SXC Pressure add variants menu (Mermaid's Lake / Shallow Water)
This adds a dialog that appears before the SXC difficulty dialog,
allowing the map to be played with low water from the start (which
is expected to make the first 30-40 turns easier, and get the map
explored faster).

The dialogs about opening sluices haven't changed, but players
will probably understand that.